refactor(cache): change size types from int64_t/size_t to uint32_t
The change improves type consistency across the cache implementation and ensures non-negative size values. The spin_mutex from TBB is also introduced for potentially better performance in concurrent scenarios.

+++ src/robot/CachedRobot.h
@@ -32,7 +32,7 @@ THE SOFTWARE.
 namespace dab::__detail__::robot {
 
 class CachedRobot : public RobotWapper<CachedRobot> {
-  static constexpr int64_t CacheSize = static_cast<int64_t>(Edge::Max) << 8;
+  static constexpr uint32_t CacheSize = static_cast<int64_t>(Edge::Max) << 8;
 
  public:
   CachedRobot() = default;
